Extract the anonymous class
